stop for now -- need to add an "unzip" intrinsic in rust!
