+++ src/python/pants/backend/jvm/register.py
@@ -9,6 +9,7 @@ from pants.backend.jvm.ossrh_publication_metadata import (
   Scm,
 )
 from pants.backend.jvm.repository import Repository as repo
+from pants.backend.jvm.rules.scalafmt import rules as scalafmt_rules
 from pants.backend.jvm.scala_artifact import ScalaArtifact
 from pants.backend.jvm.subsystems.jar_dependency_management import JarDependencyManagementSetup
 from pants.backend.jvm.subsystems.scala_platform import ScalaPlatform
@@ -228,3 +229,7 @@ def register_goals():
   task(name='test-jvm-prep-command', action=RunTestJvmPrepCommand).install('test', first=True)
   task(name='binary-jvm-prep-command', action=RunBinaryJvmPrepCommand).install('binary', first=True)
   task(name='compile-jvm-prep-command', action=RunCompileJvmPrepCommand).install('compile', first=True)
+
+
+def rules():
+  return [*scalafmt_rules()]
